pyproject.toml
[project]
name = "rag-l1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
]
